[BSE-5] zero-copy DATE32 support for arrow (#5253)
Changes the default representation of dates to be 32-bit signed integers representing the offset in days since the day of the Unix epoch (1970-01-01). This matches the in-memory Arrow representation so we can share buffers and avoid copies during IO.

* [BE-4515] [BE-4519] Use arrow format for dates internally

* [BE-4517] Modify non-serialization C++ code to handle 32-bit dates (#5208)

* [BE-4518] Date32 support in Arrow (#5227)

* Switch to new array allocation type

* Move min_row_number_filter with multiple order by columns to use 32 bit date

updated row_number_filter code

---------

+++ bodo/libs/array.py
@@ -480,7 +480,7 @@ def array_to_info_codegen(context, builder, sig, args):
         if isinstance(arr_type, DecimalArrayType):
             np_dtype = int128_type
         elif arr_type == datetime_date_array_type:
-            np_dtype = types.int64
+            np_dtype = types.int32
         elif arr_type == boolean_array_type:
             np_dtype = types.int8
         data_arr = context.make_array(types.Array(np_dtype, 1, "C"))(
@@ -1020,7 +1020,7 @@ def info_to_array_codegen(context, builder, sig, args):
         if isinstance(arr_type, DecimalArrayType):
             np_dtype = int128_type
         elif arr_type == datetime_date_array_type:
-            np_dtype = types.int64
+            np_dtype = types.int32
         elif arr_type == boolean_array_type:
             # Boolean array stores bits so we can't use boolean.
             np_dtype = types.uint8
